SKYEDEN-3234 | Move leader and config to separate classes
