+++ hermes-management/src/main/java/pl/allegro/tech/hermes/management/infrastructure/detection/InactiveTopicsDetectionScheduler.java
@@ -1,82 +1,30 @@
 package pl.allegro.tech.hermes.management.infrastructure.detection;
 
-import jakarta.annotation.PostConstruct;
-import java.util.Optional;
-import org.apache.curator.framework.CuratorFramework;
-import org.apache.curator.framework.recipes.leader.LeaderLatch;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.scheduling.annotation.Scheduled;
-import org.springframework.stereotype.Component;
-import pl.allegro.tech.hermes.infrastructure.zookeeper.ZookeeperPaths;
-import pl.allegro.tech.hermes.management.config.detection.InactiveTopicsDetectionProperties;
 import pl.allegro.tech.hermes.management.domain.detection.InactiveTopicsDetectionJob;
-import pl.allegro.tech.hermes.management.infrastructure.zookeeper.ZookeeperClient;
-import pl.allegro.tech.hermes.management.infrastructure.zookeeper.ZookeeperClientManager;
 
-@ConditionalOnProperty(prefix = "detection.inactive-topics", value = "enabled", havingValue = "true")
-@Component
-@EnableConfigurationProperties(InactiveTopicsDetectionProperties.class)
 public class InactiveTopicsDetectionScheduler {
   private final InactiveTopicsDetectionJob job;
-  private final String leaderElectionDc;
-  private final Optional<LeaderLatch> leaderLatch;
+  private final InactiveTopicsDetectionLeader leader;
 
   private static final Logger logger =
       LoggerFactory.getLogger(InactiveTopicsDetectionScheduler.class);
 
   public InactiveTopicsDetectionScheduler(
-      InactiveTopicsDetectionJob job,
-      ZookeeperClientManager zookeeperClientManager,
-      InactiveTopicsDetectionProperties inactiveTopicsDetectionProperties,
-      ZookeeperPaths zookeeperPaths) {
-    this.leaderElectionDc = inactiveTopicsDetectionProperties.leaderElectionZookeeperDc();
-    String leaderPath = zookeeperPaths.inactiveTopicsLeaderPath();
-    Optional<CuratorFramework> leaderCuratorFramework =
-        zookeeperClientManager.getClients().stream()
-            .filter(it -> it.getDatacenterName().equals(leaderElectionDc))
-            .findFirst()
-            .map(ZookeeperClient::getCuratorFramework);
-    this.leaderLatch = leaderCuratorFramework.map(it -> new LeaderLatch(it, leaderPath));
-    if (leaderLatch.isEmpty()) {
-      logLeaderZookeeperClientNotFound(leaderElectionDc);
-    } else {
-      logger.info("Found leader Zookeeper Client, leader latch created");
-    }
+      InactiveTopicsDetectionJob job, InactiveTopicsDetectionLeader leader) {
+    this.leader = leader;
     this.job = job;
-  }
-
-  @PostConstruct
-  public void startListeningForLeadership() {
-    leaderLatch.ifPresent(
-        it -> {
-          logger.info("Starting listening for leadership");
-          try {
-            it.start();
-          } catch (Exception e) {
-            throw new RuntimeException(e);
-          }
-        });
   }
 
   @Scheduled(cron = "${detection.inactive-topics.cron}")
   public void run() {
-    if (leaderLatch.isPresent()) {
-      if (leaderLatch.get().hasLeadership()) {
-        logger.info("Inactive topics detection started");
-        job.detectAndNotify();
-      } else {
-        logger.info("Inactive topics detection not started - not a leader");
-      }
+    if (leader.isLeader()) {
+      logger.info("Inactive topics detection started");
+      job.detectAndNotify();
     } else {
-      logLeaderZookeeperClientNotFound(leaderElectionDc);
+      logger.info("Inactive topics detection not started - not a leader");
     }
-  }
-
-  private void logLeaderZookeeperClientNotFound(String dc) {
-    logger.error(
-        "Cannot run inactive topics detection - no zookeeper client for datacenter={}", dc);
   }
 }
